tr_sys/requirements.txt
aiofiles==24.1.0
amqp==5.3.1
anyio==4.5.2
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
billiard==4.2.1
bleach==6.1.0
Brotli==1.1.0
celery==5.5.3
certifi==2025.4.26
channels==4.2.2
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Deprecated==1.2.18
django==4.2.23
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-cors-headers==4.4.0
django-markdownify==0.9.5
django-timezone-field==7.1
exceptiongroup==1.3.0
factory-boy==3.3.3
Faker==35.2.2
grpcio==1.70.0
h11==0.16.0
html5tagger==1.3.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
Markdown==3.7
mpmath==1.3.0
multidict==6.1.0
numpy==1.24.4
objsize==0.7.1
outcome==1.3.0.post0
packaging==25.0
pluggy==1.5.0
prompt-toolkit==3.0.51
protobuf==5.29.5
pycryptodome==3.23.0
pydantic==1.10.22
PyMySQL==1.1.1
pytest==8.3.5
pytest-django==4.11.1
pytest-asyncio
pytest-factoryboy==2.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-Levenshtein-wheels==0.13.2
PyYAML==6.0.2
reasoner-pydantic==5.0.6
requests==2.32.4
sanic==24.12.0
sanic-ext==24.12.0
sanic-routing==23.12.0
scipy==1.10.1
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
sympy==1.13.3
thrift==0.22.0
tinycss2==1.2.1
tomli==2.2.1
tracerite==1.1.1
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.2.3
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.17.2
wsproto==1.2.0
zipp==3.20.2
zstandard
git+https://github.com/biothings/biothings_annotator#egg=biothings_annotator
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-django
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-celery